requirements.txt
streamlit
py-readability-metrics
vaderSentiment
pandas
numpy
Readability
nltk 
stopwords  
Pillow
matplotlib


